chore(exactcurrency): Do not allow doubles in exactCurrency

+++ src/integration/lib/src/test/java/se/tink/backend/aggregation/nxgen/core/account/nxbuilders/modules/creditcard/CreditCardModuleTest.java
@@ -2,6 +2,7 @@ package se.tink.backend.aggregation.nxgen.core.account.nxbuilders.modules.credit
 
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigDecimal;
 import org.junit.Test;
 import se.tink.libraries.amount.ExactCurrencyAmount;
 
@@ -21,8 +22,8 @@ public class CreditCardModuleTest {
     public void missingCardNumber() {
         CreditCardModule.builder()
                 .withCardNumber(null)
-                .withBalance(ExactCurrencyAmount.of(233, "SEK"))
-                .withAvailableCredit(ExactCurrencyAmount.of(487, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(233), "SEK"))
+                .withAvailableCredit(ExactCurrencyAmount.of(BigDecimal.valueOf(487), "SEK"))
                 .withCardAlias("PRO MerVärde MasterCard")
                 .build();
     }
@@ -31,8 +32,8 @@ public class CreditCardModuleTest {
     public void invalidCardNumber() {
         CreditCardModule.builder()
                 .withCardNumber("ABC")
-                .withBalance(ExactCurrencyAmount.of(233, "SEK"))
-                .withAvailableCredit(ExactCurrencyAmount.of(487, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(233), "SEK"))
+                .withAvailableCredit(ExactCurrencyAmount.of(BigDecimal.valueOf(487), "SEK"))
                 .withCardAlias("PRO MerVärde MasterCard")
                 .build();
     }
@@ -41,8 +42,8 @@ public class CreditCardModuleTest {
     public void tooShortCardNumber() {
         CreditCardModule.builder()
                 .withCardNumber("******123")
-                .withBalance(ExactCurrencyAmount.of(233, "SEK"))
-                .withAvailableCredit(ExactCurrencyAmount.of(487, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(233), "SEK"))
+                .withAvailableCredit(ExactCurrencyAmount.of(BigDecimal.valueOf(487), "SEK"))
                 .withCardAlias("PRO MerVärde MasterCard")
                 .build();
     }
@@ -51,8 +52,8 @@ public class CreditCardModuleTest {
     public void missingCardAlias() {
         CreditCardModule.builder()
                 .withCardNumber("0000-1111-2222-3333")
-                .withBalance(ExactCurrencyAmount.of(233, "SEK"))
-                .withAvailableCredit(ExactCurrencyAmount.of(487, "SEK"))
+                .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(233), "SEK"))
+                .withAvailableCredit(ExactCurrencyAmount.of(BigDecimal.valueOf(487), "SEK"))
                 .withCardAlias(null)
                 .build();
     }
@@ -65,8 +66,8 @@ public class CreditCardModuleTest {
         CreditCardModule cardModule =
                 CreditCardModule.builder()
                         .withCardNumber(cardNumber)
-                        .withBalance(ExactCurrencyAmount.of(233, "SEK"))
-                        .withAvailableCredit(ExactCurrencyAmount.of(487, "SEK"))
+                        .withBalance(ExactCurrencyAmount.of(BigDecimal.valueOf(233), "SEK"))
+                        .withAvailableCredit(ExactCurrencyAmount.of(BigDecimal.valueOf(487), "SEK"))
                         .withCardAlias(cardAlias)
                         .build();
